server/requirements.txt
fastapi
uvicorn
python-dotenv
pdfminer.six
python-docx
openai
